+++ CopilotKit/packages/react-core/src/components/copilot-provider/copilotkit-props.tsx
@@ -34,11 +34,6 @@ export interface CopilotKitProps {
   textToSpeechUrl?: string;
 
   /**
-   * @deprecated use runtimeUrl instead
-   */
-  url?: string;
-
-  /**
    * Additional headers to be sent with the request.
    *
    * For example:
@@ -67,7 +62,7 @@ export interface CopilotKitProps {
   children: ReactNode;
 
   /**
-   * Backend only props that will be combined to body params to be sent with the request
+   * Custom properties to be sent with the request
    * For example:
    * ```js
    * {
@@ -75,7 +70,7 @@ export interface CopilotKitProps {
    * }
    * ```
    */
-  backendOnlyProps?: Record<string, any>;
+  properties?: Record<string, any>;
 
   /**
    * Indicates whether the user agent should send or receive cookies from the other domain