+++ CopilotKit/packages/runtime/src/lib/integrations/node-express/index.ts
@@ -1,0 +1,14 @@
+import { CreateCopilotRuntimeServerOptions } from "../shared";
+import { copilotRuntimeNodeHttpEndpoint } from "../node-http";
+import telemetry from "../../telemetry-client";
+
+export function copilotRuntimeNodeExpressEndpoint(options: CreateCopilotRuntimeServerOptions) {
+  telemetry.setGlobalProperties({
+    runtime: {
+      framework: "node-express",
+    },
+  });
+
+  telemetry.capture("oss.runtime.instance_created", {});
+  return copilotRuntimeNodeHttpEndpoint(options);
+}