+++ CopilotKit/packages/react-core/CHANGELOG.md
@@ -1,5 +1,72 @@
 # ui
 
+## 1.0.0
+
+### Major Changes
+
+- b6a4b6eb: V1.0 Release Candidate
+
+  - A robust new protocol between the frontend and the Copilot Runtime
+  - Support for Copilot Cloud
+  - Generative UI
+  - Support for LangChain universal tool calling
+  - OpenAI assistant API streaming
+
+- V1.0 Release
+
+  - A robust new protocol between the frontend and the Copilot Runtime
+  - Support for Copilot Cloud
+  - Generative UI
+  - Support for LangChain universal tool calling
+  - OpenAI assistant API streaming
+
+### Patch Changes
+
+- b6a4b6eb: Introduce anonymous telemetry
+- b6a4b6eb: Set default Copilot Cloud runtime URL to versioned URL (v1)
+- Updated dependencies [b6a4b6eb]
+- Updated dependencies [b6a4b6eb]
+- Updated dependencies [b6a4b6eb]
+- Updated dependencies
+  - @copilotkit/runtime-client-gql@1.0.0
+  - @copilotkit/shared@1.0.0
+
+## 1.0.0-beta.2
+
+### Patch Changes
+
+- Set default Copilot Cloud runtime URL to versioned URL (v1)
+- Updated dependencies
+  - @copilotkit/runtime-client-gql@1.0.0-beta.2
+  - @copilotkit/shared@1.0.0-beta.2
+
+## 1.0.0-beta.1
+
+### Patch Changes
+
+- Introduce anonymous telemetry
+- Updated dependencies
+  - @copilotkit/runtime-client-gql@1.0.0-beta.1
+  - @copilotkit/shared@1.0.0-beta.1
+
+## 1.0.0-beta.0
+
+### Major Changes
+
+- V1.0 Release Candidate
+
+  - A robust new protocol between the frontend and the Copilot Runtime
+  - Support for Copilot Cloud
+  - Generative UI
+  - Support for LangChain universal tool calling
+  - OpenAI assistant API streaming
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/runtime-client-gql@1.0.0-beta.0
+  - @copilotkit/shared@1.0.0-beta.0
+
 ## 0.37.0
 
 ### Minor Changes